rag_project/requirements.txt
# Core ML & DL
torch>=2.6.0
torchvision
torchaudio

# PDF parsing
pymupdf>=1.24.0

# Embeddings
sentence-transformers>=3.0.0

# Vector store
chromadb>=0.5.0

# LLM inference (GGUF models)
llama-cpp-python>=0.3.0

# Utilities
tqdm>=4.66.0
numpy>=1.26.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.29.0
ipywidgets>=8.1.0
